requirements.txt
opencv-python==4.8.1.78
pyautogui==0.9.54
pillow==10.0.1
numpy==1.24.3
keyboard==0.13.5
mss==9.0.1
